refactor(cookies): dedupe header parsing
