+++ packages/vinext/src/utils/parse-cookie.ts
@@ -1,10 +1,22 @@
 export type ParsedCookies = Record<string, string>;
 
-function decodeCookieValue(value: string): string {
+type DecodedCookieValue = { ok: true; value: string } | { ok: false };
+
+function decodeCookieValue(value: string): DecodedCookieValue {
   try {
-    return decodeURIComponent(value);
+    return { ok: true, value: decodeURIComponent(value) };
   } catch {
-    return value;
+    return { ok: false };
+  }
+}
+
+function forEachCookieHeaderPart(
+  cookieHeader: string,
+  visit: (part: string, separator: number) => void,
+): void {
+  for (const part of cookieHeader.split(/; */)) {
+    if (!part) continue;
+    visit(part, part.indexOf("="));
   }
 }
 
@@ -16,20 +28,41 @@ export function parseCookieHeader(cookieHeader: string | null | undefined): Pars
   const cookies: ParsedCookies = {};
   if (!cookieHeader) return cookies;
 
-  for (const part of cookieHeader.split(/; */)) {
-    const separator = part.indexOf("=");
-    if (separator < 0) continue;
+  forEachCookieHeaderPart(cookieHeader, (part, separator) => {
+    if (separator < 0) return;
 
     const key = part.slice(0, separator).trim();
     let value = part.slice(separator + 1).trim();
-    if (cookies[key] !== undefined) continue;
+    if (cookies[key] !== undefined) return;
 
     if (value.startsWith('"')) {
       value = value.slice(1, -1);
     }
 
-    cookies[key] = decodeCookieValue(value);
-  }
+    const decoded = decodeCookieValue(value);
+    cookies[key] = decoded.ok ? decoded.value : value;
+  });
+
+  return cookies;
+}
+
+/**
+ * Parse a Cookie header using Next.js/@edge-runtime RequestCookies semantics.
+ */
+export function parseEdgeRequestCookieHeader(cookieHeader: string): Map<string, string> {
+  const cookies = new Map<string, string>();
+
+  forEachCookieHeaderPart(cookieHeader, (part, separator) => {
+    if (separator === -1) {
+      cookies.set(part, "true");
+      return;
+    }
+
+    const decoded = decodeCookieValue(part.slice(separator + 1));
+    if (decoded.ok) {
+      cookies.set(part.slice(0, separator), decoded.value);
+    }
+  });
 
   return cookies;
 }
